typescriptize AppSetting and fix fileUpload default value
